refactor(launch-target): improve code
