+++ core/domain/src/jvmMain/kotlin/dev/koga/deeplinklauncher/util/ext/ListExt.kt
