+++ core/domain/src/jvmTest/kotlin/dev/koga/deeplinklauncher/model/FakeDevice.kt
